Testable client
