+++ test/Makefile.am
@@ -1,0 +1,7 @@
+TESTS = \
+  http_server_test.rb
+
+TEST_EXTENSIONS = .rb
+RB_LOG_COMPILER = rspec
+VERBOSE = 1
+EXTRA_DIST = $(TESTS)
