+++ test/http_server_test.rb
@@ -1,0 +1,7 @@
+#!/usr/bin/env rspec
+
+require_relative "test_helper"
+require "http-server/clients/main"
+
+describe Yast::HttpServerClient do
+end
